Run custom signal handler tests on macOS
This commit fixes a mistake where these tests were not being run on
macOS after signal handling was moved behind a `Config` flag instead of
being a crate feature. Additionally these tests are moved to their own
test file to ensure that each process either uses Mach Ports or signals
(this one uses signals which is non-default).
